# docs/contract/openapi.json
{
  "components": {
    "schemas": {
...
      "ClaimIn": {
        "additionalProperties": false,
        "properties": {
          "admission_token": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Admission Token",
            "type": "string"
          }
        },
        "required": [
          "admission_token"
        ],
        "title": "ClaimIn",
        "type": "object"
      },
      "ClaimOut": {
        "properties": {
          "allocation_id": {
            "format": "uuid",
            "title": "Allocation Id",
            "type": "string"
          },
          "confirmed_at": {
            "format": "date-time",
            "title": "Confirmed At",
            "type": "string"
          },
          "seat_no": {
            "title": "Seat No",
            "type": "integer"
          },
          "server_time": {
            "description": "ISO 8601 UTC with milliseconds, e.g. 2026-10-04T10:15:30.123Z",
            "title": "Server Time",
            "type": "string"
          }
        },
        "required": [
          "server_time",
          "allocation_id",
          "seat_no",
          "confirmed_at"
        ],
        "title": "ClaimOut",
        "type": "object"
      },
...
      "EntryIn": {
        "additionalProperties": false,
        "properties": {},
        "title": "EntryIn",
        "type": "object"
      },
      "EntryOut": {
        "properties": {
          "entry_id": {
            "format": "uuid",
            "title": "Entry Id",
            "type": "string"
          },
          "server_time": {
            "description": "ISO 8601 UTC with milliseconds, e.g. 2026-10-04T10:15:30.123Z",
            "title": "Server Time",
            "type": "string"
          },
          "status": {
            "const": "REGISTERED",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "server_time",
          "entry_id",
          "status"
        ],
        "title": "EntryOut",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "The single error envelope. Every non-2xx response has this shape.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "server_time": {
            "title": "Server Time",
            "type": "string"
          }
        },
        "required": [
          "error",
          "server_time"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "OtpRequestIn": {
        "additionalProperties": false,
        "properties": {
          "device_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Device Id",
            "type": "string"
          },
          "phone": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Phone",
            "type": "string"
          }
        },
        "required": [
          "phone",
          "device_id"
        ],
        "title": "OtpRequestIn",
        "type": "object"
      },
      "OtpRequestOut": {
        "properties": {
          "dev_otp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "SIM_MODE only; null otherwise.",
            "title": "Dev Otp"
          },
          "expires_in_s": {
            "title": "Expires In S",
            "type": "integer"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "server_time": {
            "description": "ISO 8601 UTC with milliseconds, e.g. 2026-10-04T10:15:30.123Z",
            "title": "Server Time",
            "type": "string"
          }
        },
        "required": [
          "server_time",
          "request_id",
          "expires_in_s",
          "dev_otp"
        ],
        "title": "OtpRequestOut",
        "type": "object"
      },
      "OtpVerifyIn": {
        "additionalProperties": false,
        "properties": {
          "device_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Device Id",
            "type": "string"
          },
          "otp": {
            "maxLength": 16,
            "minLength": 1,
            "title": "Otp",
            "type": "string"
          },
          "request_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "otp",
          "device_id"
        ],
        "title": "OtpVerifyIn",
        "type": "object"
      },
      "OtpVerifyOut": {
        "properties": {
          "server_time": {
            "description": "ISO 8601 UTC with milliseconds, e.g. 2026-10-04T10:15:30.123Z",
            "title": "Server Time",
            "type": "string"
          },
          "session_token": {
            "title": "Session Token",
            "type": "string"
          },
          "user_public_id": {
            "title": "User Public Id",
            "type": "string"
          }
        },
        "required": [
          "server_time",
          "session_token",
          "user_public_id"
        ],
        "title": "OtpVerifyOut",
        "type": "object"
      },
...
      "StepUpIn": {
        "additionalProperties": false,
        "properties": {
          "otp": {
            "maxLength": 16,
            "minLength": 1,
            "title": "Otp",
            "type": "string"
          }
        },
        "required": [
          "otp"
        ],
        "title": "StepUpIn",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AdminKey": {
        "in": "header",
        "name": "X-Admin-Key",
        "type": "apiKey"
      },
      "SessionBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "SessionCookie": {
        "in": "cookie",
        "name": "fd_session",
        "type": "apiKey"
      },
      "SimKey": {
        "in": "header",
        "name": "X-Sim-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Fair Drop public, admin and simulator API. Errors: one envelope.",
    "title": "Fair Drop API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/otp/request": {
      "post": {
        "description": "Same phone within 30 s returns the same request_id. dev_otp only when SIM_MODE.",
        "operationId": "requestOtp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_PHONE | VALIDATION_ERROR"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "OTP_THROTTLED | RATE_LIMITED"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "SERVICE_UNAVAILABLE"
          }
        },
        "summary": "Request Otp",
        "tags": [
          "public"
        ]
      }
    },
    "/api/auth/otp/verify": {
      "post": {
        "description": "Sets the fd_session cookie. Re-verify returns the existing session for the same device.",
        "operationId": "verifyOtp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpVerifyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "VALIDATION_ERROR"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "OTP_INVALID"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "OTP_EXPIRED"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RATE_LIMITED"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "SERVICE_UNAVAILABLE"
          }
        },
        "summary": "Verify Otp",
        "tags": [
          "public"
        ]
      }
    },
    "/api/drops/{drop_id}": {
      "get": {
        "description": "Public drop state. Read-only; cacheable for 1 s.",
        "operationId": "getDrop",
        "parameters": [
          {
            "in": "path",
            "name": "drop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Drop Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DropOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "VALIDATION_ERROR"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "NOT_FOUND"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "SERVICE_UNAVAILABLE"
          }
        },
        "summary": "Get Drop",
        "tags": [
          "public"
        ]
      }
    },
    "/api/drops/{drop_id}/claim": {
      "post": {
        "description": "Requires Idempotency-Key. Same key or same entry returns the same 200.",
        "operationId": "claim",
        "parameters": [
          {
            "in": "path",
            "name": "drop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Drop Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "IDEMPOTENCY_KEY_MISSING | VALIDATION_ERROR"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "TOKEN_INVALID"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "NOT_OFFERED"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "OFFER_EXPIRED | SOLD_OUT"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "IDEMPOTENCY_KEY_REUSED"
          },
          "423": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "STEP_UP_REQUIRED"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RATE_LIMITED"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "SERVICE_UNAVAILABLE"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionBearer": []
          }
        ],
        "summary": "Claim",
        "tags": [
          "public"
        ]
      }
    },
    "/api/drops/{drop_id}/draw-proof": {
      "get": {
        "description": "Addition: the proof plus the sorted eligible public ids and the ranked order, so anyone can\nre-run the draw (docs/contract/draw.md). Only after the draw.",
        "operationId": "publicDrawProof",
        "parameters": [
          {
            "in": "path",
            "name": "drop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Drop Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrawProofOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "VALIDATION_ERROR"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "NOT_FOUND"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_TRANSITION"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "SERVICE_UNAVAILABLE"
          }
        },
        "summary": "Public Draw Proof",
        "tags": [
          "public"
        ]
      }
    },
    "/api/drops/{drop_id}/entries": {
      "post": {
        "description": "Idempotent by UNIQUE(drop, user). Idempotency-Key is optional here.",
        "operationId": "createEntry",
        "parameters": [
          {
            "in": "path",
            "name": "drop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Drop Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntryIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryOut"
                }
              }
            },
            "description": "Repeat call: same body as the 201."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "VALIDATION_ERROR"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "UNAUTHENTICATED"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "WINDOW_CLOSED | WINDOW_NOT_OPEN"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RATE_LIMITED"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "SERVICE_UNAVAILABLE"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionBearer": []
          }
        ],
        "summary": "Create Entry",
        "tags": [
          "public"
        ]
      }
    },
    "/api/drops/{drop_id}/me": {
      "get": {
        "description": "The UI's single source of truth. Read-only.",
        "operationId": "getMe",
        "parameters": [
          {
            "in": "path",
            "name": "drop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Drop Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "VALIDATION_ERROR"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "UNAUTHENTICATED"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RATE_LIMITED"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "SERVICE_UNAVAILABLE"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionBearer": []
          }
        ],
        "summary": "Get Me",
        "tags": [
          "public"
        ]
      }
    },
    "/api/drops/{drop_id}/step-up": {
      "post": {
        "description": "Requires Idempotency-Key. A repeat after success returns 200.",
        "operationId": "stepUp",
        "parameters": [
          {
            "in": "path",
            "name": "drop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Drop Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepUpIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepUpOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "IDEMPOTENCY_KEY_MISSING | VALIDATION_ERROR"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "OTP_INVALID"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "OFFER_EXPIRED"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "SERVICE_UNAVAILABLE"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionBearer": []
          }
        ],
        "summary": "Step Up",
        "tags": [
          "public"
        ]
      }
    },
...
  }
}